requirements.txt
Faker==18.11.2
python-docx==1.1.0
openpyxl==3.1.2
fpdf==1.7.2
PyPDF2==3.0.1
python-magic==0.4.27
chardet==5.2.0
rarfile==4.0
py7zr==0.20.5